+++ src/app/common/core/services/services.module.ts
@@ -1,0 +1,14 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+
+import { TransactionsService } from './transactions.service';
+
+@NgModule({
+  imports: [
+    CommonModule
+  ],
+  providers: [
+    TransactionsService
+  ]
+})
+export class ServicesModule { }
