--- src/app/features/main/features/transactions-list/services/transactions.service.ts
@@ -1,18 +1,0 @@
-import { Injectable } from '@angular/core';
-
-import { HttpClient } from '@angular/common/http';
-
-@Injectable()
-export class TransactionsService {
-
-constructor(private http: HttpClient) { }
-
-  getTransactions(data = {}) {
-    return this.http.post(`transaction`, data);
-  }
-
-  removeExpense(id) {
-    return this.http.delete(`transaction/${id}`);
-  }
-
-}
